Keep .git dir in case of blank transformation on same folder #292

+++ butterfly-core/src/main/java/com/paypal/butterfly/core/TransformationEngine.java
@@ -567,6 +567,7 @@ class TransformationEngine {
                 try {
                     FileUtils.copyDirectory(application.getFolder(), baselineAppFolder);
                     for (File file : application.getFolder().listFiles()) {
+                        if (file.getName().equals(".git")) continue;
                         FileUtils.forceDelete(file);
                     }
                 } catch (IOException e) {
